+++ lib/src/widgets/chat_view_inherited_widget.dart
@@ -10,8 +10,10 @@ class ChatViewInheritedWidget extends InheritedWidget {
     required this.featureActiveConfig,
     required this.chatController,
     required this.currentUser,
+    this.profileCircleConfiguration,
   }) : super(key: key, child: child);
   final FeatureActiveConfig featureActiveConfig;
+  final ProfileCircleConfiguration? profileCircleConfiguration;
   final ChatController chatController;
   final ChatUser currentUser;