+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [2.0.0] (Unreleased)
 
+* **Feat**: [156](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/156) Added
+  default avatar, error builder for asset, network and base64 profile image and
+  cached_network_image for network images.
 * **Breaking**: [173](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/173) Added
   callback to sort message in chat.
 * **Fix**: [181](https://github.com/SimformSolutionsPvtLtd/flutter_chatview/pull/181) Removed